Move to multi.c
