+++ src/redisraft.h
@@ -331,7 +331,7 @@ typedef struct RedisRaftCtx {
     struct RaftReq *transfer_req;                /* RaftReq if a leader transfer is in progress */
     RedisModuleCommandFilter *registered_filter; /* Command filter is used for intercepting redis commands */
     struct ShardingInfo *sharding_info;          /* Information about sharding, when cluster mode is enabled */
-    RedisModuleDict *multiClientState;           /* A dict that tracks multi state of the clients */
+    RedisModuleDict *multi_client_state;         /* A dict that tracks multi state of the clients */
 
     /* General stats */
     unsigned long client_attached_entries;       /* Number of log entries attached to user connections */
@@ -856,4 +856,9 @@ const CommandSpec *CommandSpecGet(const RedisModuleString *cmd);
 /* sort.c */
 void handleSort(RedisModuleCtx *ctx, RedisModuleString **argv, int argc);
 
+/* multi.c */
+void MultiInitClientState(RedisRaftCtx *rr);
+uint64_t MultiClientStateCount(RedisRaftCtx *rr);
+void MultiFreeClientState(RedisRaftCtx *rr, unsigned long long client_id);
+bool MultiHandleCommand(RedisRaftCtx *rr, RedisModuleCtx *ctx, RaftRedisCommandArray *cmds);
 #endif  /* _REDISRAFT_H */
